+++ .oldindex.js
@@ -1,82 +1,57 @@
-import { Chain } from "./functions/src/blockchain-core/chain.js";
-import { getGenesisBlock, getLastBlock, sendBlockToServer, sendChainToServer, sendTransactionToServer, getBlockById } from "./functions/src/blockchain-core/database_logic.js";
-import { Wallet } from "./functions/src/blockchain-core/wallet.js";
+import { Chain } from "./functions/src/blockchain-core/chain";
+import { getGenesisBlock, getLastBlock, sendBlockToServer, sendChainToServer, sendTransactionToServer, getBlockById } from "./functions/src/blockchain-core/database_logic";
+import { Wallet } from "./functions/src/blockchain-core/wallet";
 import * as readline from 'readline';
-import { startNode, sendMessage, sendTransaction, setHandleMessageCallback, setOnNodeDisconnectedCallback, getCurrentNodes } from './libp2p-conexion/streamer.js';
-
-
-
-const thisWallet = new Wallet()
-
+import { startNode, sendMessage, setHandleMessageCallback, setOnNodeDisconnectedCallback } from './libp2p-conexion/streamer';
+const wallets = [];
 export const initializeGenesisBlock = async () => {
     console.log("Initializing genesis block OR retrieve of the last block from chaindb");
-
-
     try {
-
         const primaryBlock = Chain.instance.lastBlock;
-        console.log(primaryBlock)
+        console.log(primaryBlock);
         await sendBlockToServer(primaryBlock);
-
         await sendTransactionToServer(Chain.instance.lastBlock.transaction);
-
         await sendChainToServer(Chain.instance.chain);
-
         console.log("Initialization complete");
-    } catch (error) {
+    }
+    catch (error) {
         console.error("Initialization error:", error);
     }
-}
-
-
+};
 export const setSecondaryBlockInitialize = async () => {
     console.log("Initializing secondary block...");
     const sender = new Wallet();
     const recipient = new Wallet();
     await sender.sendMoney(50, recipient.publicKey);
     await sendChainToServer(Chain.instance.chain);
-}
-
+};
 export const sendGenericTransaction = async () => {
     console.log("Sending a generic transaction...");
-
     // Generate random values for the transaction
     const amount = Math.floor(Math.random() * 100); // Random amount (0 to 99)
     const sender = new Wallet();
     const recipient = new Wallet();
-
     // Send the random transaction
     await sender.sendMoney(amount, recipient.publicKey);
-
-
     await sendChainToServer(Chain.instance.chain);
-
     console.log("Generic transaction sent successfully!");
 };
-
-
 export async function getGenesisBlockFromDatabase() {
     const genesisBlock = await getGenesisBlock();
     console.log("Get genesis block from database:", genesisBlock);
 }
-
 export async function getLastBlockFromDatabase() {
     const lastBlock = await getLastBlock();
     console.log("Get last block from database blockchain:", lastBlock);
 }
-
 export async function getBlockByIdFromDatabase(id) {
     const idblock = await getBlockById(id);
     console.log(`Block with id ${id} from database blockchain:`, idblock);
 }
-
-
-
 export const getFunctionsFromDataBase = async () => {
     await getGenesisBlockFromDatabase();
     await getLastBlockFromDatabase();
-}
-
+};
 function formatAndSummarizePublicKey(publicKey) {
     const keyLines = publicKey.split('\n');
     const firstLine = keyLines[0]; // Include the first line "-----BEGIN PUBLIC KEY-----"
@@ -85,7 +60,6 @@ function formatAndSummarizePublicKey(publicKey) {
     const lastChars = lastLine.slice(-4);
     return `${firstLine}\n  ${firstChars}...${lastChars}\n${lastLine}`;
 }
-
 function displayBlockParts(block) {
     console.log("Block Details:");
     console.log(`Index: ${block.index}`);
@@ -104,60 +78,42 @@ function clearConsole() {
     if (isWindows) {
         // For Windows
         process.stdout.write("\x1Bc");
-    } else {
+    }
+    else {
         // For other operating systems
         console.clear();
     }
 }
-
 setHandleMessageCallback((message) => {
     console.log("Received message from libp2p:", message);
     renderMenu();
 });
-
 setOnNodeDisconnectedCallback(() => {
     console.log("A node has disconnected. Disabling the menu.");
 });
-
-const renderWallet = (thisWallet) => {
-    // Clear the terminal screen (if needed)
-    console.log("Welcome!")
-    console.log("\nWallet Information:");
-    console.log("Current Money: ", thisWallet.money);
-};
-
 const renderMenu = () => {
     // Clear the terminal screen
-
     console.log("\nMenu:");
     console.log("1. See last block from chain");
     console.log("2. See last block from the database");
     console.log("3. Get genesis block from database");
     console.log("4. Send a generic transaction");
     console.log("5. Get block by ID");
-    console.log("6. Enviar mensaje a otros nodos (por defecto broadcast)");
-    console.log("7. Enviar transacción a otro nodo(por defecto unico otro nodo o broacast):");
-    console.log("8. Exit");
+    console.log("6. <test> ");
+    console.log("7. Exit");
     // Display received messages
-
-
 };
-
 const main = async () => {
     try {
         await Chain.instance.initializeChain();
         await sendChainToServer(Chain.instance.chain);
-        console.log("Initialization complete, WAITING for enough nodes...");
-        const response = await startNode()
-
+        console.log("Initialization complete");
+        const response = await startNode();
         console.clear();
         renderMenu();
         while (true) {
-
-            renderWallet(thisWallet)
             renderMenu();
-            const choice = await getUserInput("Choose an option (1-8): ");
-
+            const choice = await getUserInput("Choose an option (1-5): ");
             switch (choice) {
                 case '1':
                     // Action to see the last block from the chain
@@ -167,10 +123,10 @@ const main = async () => {
                     //displayBlockParts(lastBlockFromChain)
                     break;
                 case '2':
-                    await getLastBlockFromDatabase()
+                    await getLastBlockFromDatabase();
                     break;
                 case '3':
-                    await getGenesisBlockFromDatabase()
+                    await getGenesisBlockFromDatabase();
                     break;
                 case '4':
                     await sendGenericTransaction(); // Use 'await' here
@@ -180,29 +136,20 @@ const main = async () => {
                     const userInput = await getUserInput("Enter the block ID: ");
                     const blockId = String(userInput); // Ensure blockId is a string
                     await getBlockByIdFromDatabase(blockId);
-                    break;
                 case '6':
                     const message = await getUserInput('Enter the message to send: ');
-                    await sendMessage(message)
+                    await sendMessage(message);
                     console.log('Message sent.');
                     break;
                 case '7':
-                    const currentnodes = getCurrentNodes()
-                    console.log("Choose a node number:" + currentnodes)
-                    const userNodeInput = await getUserInput("Enter the node ID: ");
-                    const userMoneyToSend = await getUserInput("Enter amount of money:")
-                    const sended = await sendTransaction(userNodeInput, userMoneyToSend)
-                    if (sended) {
-                        console.log("Successfull Transacction")
-                    }
-                case '8':
                     console.log("Exiting the menu.");
                     process.exit(0);
                 default:
                     console.log("Invalid choice. Please select an option from 1 to 5.");
             }
         }
-    } catch (error) {
+    }
+    catch (error) {
         console.error("Initialization error:", error);
     }
 };
@@ -211,7 +158,6 @@ async function getUserInput(question) {
         input: process.stdin,
         output: process.stdout,
     });
-
     return new Promise((resolve) => {
         rl.question(question, (answer) => {
             rl.close();
