+++ docker-compose.yml
@@ -1,0 +1,10 @@
+version: '3'
+services:
+  blockchain:
+    build:
+      context: .
+      dockerfile: Dockerfile
+    ports:
+      - "3000:3000"
+    volumes:
+      - .:/usr/src/app
