+++ package.json
@@ -5,7 +5,9 @@
   "exports": "./start.js",
   "type": "module",
   "scripts": {
-    "start": "ts-node index.ts",
+    "start": "node index.js",
+    "startDocker" : "npx ts-node ./index.ts",
+     "build": "tsc",
     "p2p": "node libptest.js",
     "p2phttp": "node simpleMainNodelibp2p.js",
     "p2phttp2": "node l2.js",
