--- tsconfig.json
@@ -1,17 +1,0 @@
-{
-    "compilerOptions": {
-        "lib": ["ES2022", "DOM"],
-        "module": "ES2020",
-        "target": "ES2022",
-        "types": ["node"],
-        "strict": false,
-        "strictFunctionTypes": false,
-        "esModuleInterop": true,
-        "skipLibCheck": true,
-        "moduleResolution":"node",
-    },
-    "ts-node": {
-        "esm": true,
-        "experimentalSpecifierResolution": "node"
-      }
-}
